Fix eternal MSVC build errors

+++ toolkit/include/kit/multiprocessing/foreach.hpp
@@ -7,14 +7,14 @@ KIT_NAMESPACE_BEGIN
 template <typename TManager, std::random_access_iterator It1, std::output_iterator<Ref<ITask>> It2, typename Callable,
           typename... Args>
     requires std::is_base_of_v<TaskManager, TManager>
-void ForEach(TManager &p_Manager, It1 p_First, It1 p_Last, It2 p_Dest, const u32 p_Tasks, Callable &&p_Callable,
+void ForEach(TManager &p_Manager, It1 p_First, It1 p_Last, It2 p_Dest, const usize p_Tasks, Callable &&p_Callable,
              Args &&...p_Args)
 {
     const usize size = std::distance(p_First, p_Last);
-    u32 start = 0;
-    for (u32 i = 0; i < p_Tasks; ++i)
+    usize start = 0;
+    for (usize i = 0; i < p_Tasks; ++i)
     {
-        const u32 end = (i + 1) * size / p_Tasks;
+        const usize end = (i + 1) * size / p_Tasks;
         KIT_ASSERT(end <= size, "Partition exceeds container size");
         if (end > start)
         {
